Fix error in sonarcloud MU-209
