+++ src/src/main/java/com/hcmus/mentor/backend/controller/payload/response/HomePageResponse.java
@@ -1,7 +1,7 @@
 package com.hcmus.mentor.backend.controller.payload.response;
 
 import com.hcmus.mentor.backend.controller.payload.response.groups.GroupHomepageResponse;
-import com.hcmus.mentor.backend.service.impl.EventServiceImpl;
+import com.hcmus.mentor.backend.service.dto.EventDto;
 
 import java.util.List;
 
@@ -13,7 +13,7 @@ import org.springframework.data.domain.Slice;
 @NoArgsConstructor
 public class HomePageResponse {
 
-    List<EventServiceImpl.Event> events;
+    List<EventDto> events;
     List<GroupHomepageResponse> pinnedGroups;
     Slice<GroupHomepageResponse> groups;
 }
